+++ Match-Api/src/main/java/com/example/matchapi/admin/donation/convertor/AdminDonationConvertor.java
@@ -1,0 +1,26 @@
+package com.example.matchapi.admin.donation.convertor;
+
+import com.example.matchapi.donation.dto.DonationRes;
+import com.example.matchcommon.annotation.Convertor;
+import com.example.matchdomain.donation.entity.DonationUser;
+
+@Convertor
+public class AdminDonationConvertor {
+    public DonationRes.DonationDetail getDonationDetail(DonationUser donationUser) {
+        return DonationRes.DonationDetail
+                .builder()
+                .donationId(donationUser.getId())
+                .userId(donationUser.getUserId())
+                .name(donationUser.getUser().getName())
+                .email(donationUser.getUser().getEmail())
+                .phoneNumber(donationUser.getUser().getPhoneNumber())
+                .amount(donationUser.getPrice())
+                .inherenceName(donationUser.getInherenceName())
+                .inherenceNumber(donationUser.getInherenceNumber())
+                .payMethod(donationUser.getPayMethod().getValue())
+                .donationStatus(donationUser.getDonationStatus())
+                .regularStatus(donationUser.getRegularStatus().getName())
+                .donationDate(donationUser.getCreatedAt().toString())
+                .build();
+    }
+}
